combined loadtest and graphqlloadtest into one
